test_requirements.txt
coverage==4.0.3
django-crispy-forms==1.5.2
django-discover-runner==1.0
factory-boy==2.6.0
colour_runner==0.0.4
dj_database_url==0.3.0
psycopg2==2.5.3
